fix (or baseline) psalm violations
