+++ psalm-baseline.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<files psalm-version="5.26.1@d747f6500b38ac4f7dfc5edbcae6e4b637d7add0">
+<files psalm-version="6.10.1@f9fd6bc117e9ce1e854c2ed6777e7135aaa4966b">
   <file src="src/Propel/Common/Config/ConfigurationManager.php">
     <UnsupportedPropertyReferenceUsage>
       <code><![CDATA[$configSection = &$this->config[$section]]]></code>
@@ -14,50 +14,6 @@
     <UnsupportedPropertyReferenceUsage>
       <code><![CDATA[$existingNames = &self::$insertedAggregationNames[$tableName]]]></code>
     </UnsupportedPropertyReferenceUsage>
-  </file>
-  <file src="src/Propel/Generator/Behavior/ConcreteInheritance/ConcreteInheritanceBehavior.php">
-    <UndefinedPropertyAssignment>
-      <code><![CDATA[$fk->isParentChild]]></code>
-    </UndefinedPropertyAssignment>
-  </file>
-  <file src="src/Propel/Generator/Behavior/Versionable/VersionableBehavior.php">
-    <UndefinedPropertyAssignment>
-      <code><![CDATA[$syncedTable->isVersionTable]]></code>
-    </UndefinedPropertyAssignment>
-  </file>
-  <file src="src/Propel/Generator/Builder/Om/AbstractObjectBuilder.php">
-    <UndefinedMethod>
-      <code><![CDATA[addAddArrayElement]]></code>
-      <code><![CDATA[addAddArrayElement]]></code>
-      <code><![CDATA[addArrayAccessor]]></code>
-      <code><![CDATA[addArrayMutator]]></code>
-      <code><![CDATA[addBooleanAccessor]]></code>
-      <code><![CDATA[addBooleanMutator]]></code>
-      <code><![CDATA[addDefaultAccessor]]></code>
-      <code><![CDATA[addDefaultAccessor]]></code>
-      <code><![CDATA[addDefaultMutator]]></code>
-      <code><![CDATA[addEnumAccessor]]></code>
-      <code><![CDATA[addEnumMutator]]></code>
-      <code><![CDATA[addHasArrayElement]]></code>
-      <code><![CDATA[addHasArrayElement]]></code>
-      <code><![CDATA[addJsonAccessor]]></code>
-      <code><![CDATA[addJsonMutator]]></code>
-      <code><![CDATA[addLazyLoader]]></code>
-      <code><![CDATA[addLobMutator]]></code>
-      <code><![CDATA[addObjectAccessor]]></code>
-      <code><![CDATA[addObjectMutator]]></code>
-      <code><![CDATA[addRemoveArrayElement]]></code>
-      <code><![CDATA[addRemoveArrayElement]]></code>
-      <code><![CDATA[addSetAccessor]]></code>
-      <code><![CDATA[addSetMutator]]></code>
-      <code><![CDATA[addTemporalAccessor]]></code>
-      <code><![CDATA[addTemporalMutator]]></code>
-    </UndefinedMethod>
-  </file>
-  <file src="src/Propel/Generator/Builder/Om/ObjectBuilder/CrossRelationPartial.php">
-    <DuplicateArrayKey>
-      <code><![CDATA[$additionalGetterNames]]></code>
-    </DuplicateArrayKey>
   </file>
   <file src="src/Propel/Generator/Builder/Util/SchemaReader.php">
     <UnsupportedPropertyReferenceUsage>
@@ -81,17 +37,14 @@
       <code><![CDATA[$this->columnObjects]]></code>
     </InvalidPropertyAssignmentValue>
   </file>
-  <file src="src/Propel/Runtime/ActiveQuery/Criteria.php">
-    <UndefinedFunction>
-      <code><![CDATA[trigger_deprecation('Propel', '2.0', "Method $name should not be used anymore, see DeprecatedCriteriaMethods::$name how to replace it.")]]></code>
-    </UndefinedFunction>
+  <file src="src/Propel/Generator/Reverse/OracleSchemaParser.php">
+    <InvalidArrayOffset>
+      <code><![CDATA[$row[0]]]></code>
+    </InvalidArrayOffset>
   </file>
   <file src="src/Propel/Runtime/ActiveQuery/ModelCriteria.php">
     <UndefinedMethod>
       <code><![CDATA[addSelectColumns]]></code>
-      <code><![CDATA[getTableMap]]></code>
-      <code><![CDATA[getTableMap]]></code>
-      <code><![CDATA[getTableMap]]></code>
     </UndefinedMethod>
   </file>
   <file src="src/Propel/Runtime/Adapter/Pdo/MssqlAdapter.php">
@@ -111,6 +64,9 @@
     </UndefinedConstant>
   </file>
   <file src="src/Propel/Runtime/Formatter/AbstractFormatterWithHydration.php">
+    <NonVariableReferenceReturn>
+      <code><![CDATA[$this->alreadyHydratedObjects[$this->class][$mainKey]]]></code>
+    </NonVariableReferenceReturn>
     <UnsupportedPropertyReferenceUsage>
       <code><![CDATA[$arrayToAugment = &$this->alreadyHydratedObjects[$this->class][$mainKey]]]></code>
     </UnsupportedPropertyReferenceUsage>
@@ -120,11 +76,6 @@
       <code><![CDATA[populateObject]]></code>
       <code><![CDATA[populateObject]]></code>
     </UndefinedMethod>
-  </file>
-  <file src="src/Propel/Runtime/ServiceContainer/StandardServiceContainer.php">
-    <UndefinedClass>
-      <code><![CDATA[RotatingFileHandler]]></code>
-    </UndefinedClass>
   </file>
   <file src="src/Propel/Runtime/Util/Profiler.php">
     <InvalidArrayOffset>
