+++ psalm.xml
@@ -17,28 +17,8 @@
 
     <issueHandlers>
         <RedundantConditionGivenDocblockType errorLevel="suppress"/>
-        <DocblockTypeContradiction errorLevel="suppress"/>
         <MissingClosureParamType errorLevel="suppress"/>
         <MissingClosureReturnType errorLevel="suppress"/>
         <PropertyNotSetInConstructor errorLevel="suppress"/>
-        <UnresolvableInclude errorLevel="suppress"/>
-        <ImplementedReturnTypeMismatch errorLevel="suppress"/>
-        <ImplicitToStringCast errorLevel="suppress"/>
-        <UndefinedMagicMethod errorLevel="suppress"/>
-        <UndefinedMethod>
-            <errorLevel type="suppress">
-                <referencedMethod name="Symfony\Component\Config\Definition\Builder\TreeBuilder::root"/>
-            </errorLevel>
-        </UndefinedMethod>
-        <!-- Workaround for https://github.com/vimeo/psalm/issues/7026 -->
-        <ReservedWord errorLevel="suppress">
-            <errorLevel type="suppress">
-                <file name="src/Propel/Common/Config/ConfigurationManager.php" />
-                <file name="src/Propel/Common/Config/Loader/YamlFileLoader.php" />
-                <file name="src/Propel/Common/Config/PropelConfiguration.php" />
-                <file name="src/Propel/Generator/Behavior/Validate/ValidateBehavior.php" />
-                <file name="src/Propel/Runtime/Parser/YamlParser.php" />
-            </errorLevel>
-        </ReservedWord>
     </issueHandlers>
 </psalm>
