+++ src/Propel/Generator/Builder/Om/AbstractObjectBuilder.php
@@ -8,8 +8,6 @@
 
 namespace Propel\Generator\Builder\Om;
 
-use Propel\Generator\Model\PropelTypes;
-
 /**
  * Base class for object-building classes.
  *
@@ -21,105 +19,6 @@ use Propel\Generator\Model\PropelTypes;
  */
 abstract class AbstractObjectBuilder extends AbstractOMBuilder
 {
-    /**
-     * Adds the getter methods for the column values.
-     * This is here because it is probably generic enough to apply to templates being generated
-     * in different PHP versions.
-     *
-     * @param string $script The script will be modified in this method.
-     *
-     * @return void
-     */
-    protected function addColumnAccessorMethods(string &$script): void
-    {
-        $table = $this->getTable();
-
-        foreach ($table->getColumns() as $col) {
-            $type = $col->getType();
-            // if they're not using the DateTime class then we will generate "compatibility" accessor method
-            if (
-                $type === PropelTypes::DATE
-                || $type === PropelTypes::DATETIME
-                || $type === PropelTypes::TIME
-                || $type === PropelTypes::TIMESTAMP
-            ) {
-                $this->addTemporalAccessor($script, $col);
-            } elseif ($type === PropelTypes::OBJECT) {
-                $this->addObjectAccessor($script, $col);
-            } elseif ($type === PropelTypes::PHP_ARRAY) {
-                $this->addArrayAccessor($script, $col);
-                if ($col->isNamePlural()) {
-                    $this->addHasArrayElement($script, $col);
-                }
-            } elseif ($type === PropelTypes::JSON) {
-                $this->addJsonAccessor($script, $col);
-            } elseif ($col->isEnumType()) {
-                $this->addEnumAccessor($script, $col);
-            } elseif ($col->isSetType()) {
-                $this->addSetAccessor($script, $col);
-                if ($col->isNamePlural()) {
-                    $this->addHasArrayElement($script, $col);
-                }
-            } elseif ($col->isBooleanType()) {
-                $this->addDefaultAccessor($script, $col);
-                $this->addBooleanAccessor($script, $col);
-            } else {
-                $this->addDefaultAccessor($script, $col);
-            }
-
-            if ($col->isLazyLoad()) {
-                $this->addLazyLoader($script, $col);
-            }
-        }
-    }
-
-    /**
-     * Adds the mutator (setter) methods for setting column values.
-     * This is here because it is probably generic enough to apply to templates being generated
-     * in different PHP versions.
-     *
-     * @param string $script The script will be modified in this method.
-     *
-     * @return void
-     */
-    protected function addColumnMutatorMethods(string &$script): void
-    {
-        foreach ($this->getTable()->getColumns() as $col) {
-            if ($col->getType() === PropelTypes::OBJECT) {
-                $this->addObjectMutator($script, $col);
-            } elseif ($col->isLobType()) {
-                $this->addLobMutator($script, $col);
-            } elseif (
-                $col->getType() === PropelTypes::DATE
-                || $col->getType() === PropelTypes::DATETIME
-                || $col->getType() === PropelTypes::TIME
-                || $col->getType() === PropelTypes::TIMESTAMP
-            ) {
-                $this->addTemporalMutator($script, $col);
-            } elseif ($col->getType() === PropelTypes::PHP_ARRAY) {
-                $this->addArrayMutator($script, $col);
-                if ($col->isNamePlural()) {
-                    $this->addAddArrayElement($script, $col);
-                    $this->addRemoveArrayElement($script, $col);
-                }
-            } elseif ($col->getType() === PropelTypes::JSON) {
-                $this->addJsonMutator($script, $col);
-            } elseif ($col->isEnumType()) {
-                $this->addEnumMutator($script, $col);
-            } elseif ($col->isSetType()) {
-                $this->addSetMutator($script, $col);
-                if ($col->isNamePlural()) {
-                    $this->addAddArrayElement($script, $col);
-                    $this->addRemoveArrayElement($script, $col);
-                }
-            } elseif ($col->isBooleanType()) {
-                $this->addBooleanMutator($script, $col);
-            } else {
-                $this->addDefaultMutator($script, $col);
-            }
-        }
-    }
-
     /**
      * Gets the baseClass path if specified for table/db.
      *
